feat:payout implement payout

+++ database/migrations/2026_08_19_015820_create_transactions_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
 
             $table->string('direction');
 
-            $table->unsignedBigInteger('amount');
+            $table->decimal('amount', 15, 2);
 
             $table->text('description')
                 ->nullable();
