+++ database/migrations/2026_08_26_203031_create_payouts_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
                 ->constrained()
                 ->restrictOnDelete();
 
-            $table->unsignedBigInteger('amount');
+            $table->decimal('amount', 15, 2);
 
             $table->string('status');
